Moves ContactCard to separate directory
